functions/requirements.txt
firebase-functions>=0.1.0,<0.2
firebase-admin==6.6.0
google-cloud-firestore==2.19.0
google-cloud-aiplatform==1.71.1
google-cloud-storage==2.18.2
python-dotenv==1.0.1
functions-framework==3.9.2
Flask==3.0.3
